# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ResumeAI - AI-powered resume tailoring and cover letter generation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Resume {
  id: number;
  userId: string;
  title: string;
  content: string;
  /** @nullable */
  tag: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateResumeBody {
  title: string;
  content: string;
  /** @nullable */
  tag?: string | null;
}

export interface UpdateResumeBody {
  title?: string;
  content?: string;
  /** @nullable */
  tag?: string | null;
}

export interface Analysis {
  id: number;
  userId: string;
  /** @nullable */
  resumeId: number | null;
  jobTitle: string;
  /** @nullable */
  companyName: string | null;
  jobDescription: string;
  resumeContent: string;
  matchScore: number;
  tailoredResume: string;
  coverLetter: string;
  keywordsMatched: string[];
  keywordsMissing: string[];
  interviewQuestions?: string[];
  status?: string;
  createdAt: string;
}

export interface CreateAnalysisBody {
  /** @nullable */
  resumeId?: number | null;
  jobTitle: string;
  /** @nullable */
  companyName?: string | null;
  jobDescription: string;
  resumeContent: string;
}

export interface DashboardStats {
  totalResumes: number;
  totalAnalyses: number;
  averageMatchScore: number;
  plan: string;
  usageCount: number;
  usageLimit: number;
}

export interface Review {
  id: number;
  name: string;
  role: string;
  text: string;
  score: number;
  createdAt: string;
  profileImageUrl?: string;
}

export interface CreateReviewBody {
  role: string;
  text: string;
  score: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type UpdateUserSettingsBody = {
  firstName?: string;
  lastName?: string;
};

export type ExtractResumeTextBody = {
  file?: Blob;
};

export type ExtractResumeText200 = {
  text?: string;
};

export type ScrapeJobDescriptionBody = {
  url?: string;
};

export type ScrapeJobDescription200 = {
  text?: string;
};

export type GetRecentAnalysesParams = {
  limit?: number;
};