# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ResumeAI - AI-powered resume tailoring and cover letter generation
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes for the authenticated user
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
        "401":
          description: Not authenticated

    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
  /resumes/extract:
    post:
      operationId: extractResumeText
      tags: [resumes]
      summary: Extract plain text from a resume file (PDF or Docx)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Text extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
        "400":
          description: Invalid file or no text found
        "500":
          description: Extraction failed
...
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Resume not found

    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResumeBody"
      responses:
        "200":
          description: Resume updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Resume not found

    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Resume deleted
        "404":
          description: Resume not found
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List analyses for the authenticated user
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
        "401":
          description: Not authenticated

    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "403":
          description: Free plan limit reached
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get an analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Analysis not found

    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Analysis deleted
        "404":
          description: Analysis not found
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary stats for the authenticated user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Not authenticated
  /dashboard/recent-analyses:
    get:
      operationId: getRecentAnalyses
      tags: [dashboard]
      summary: Get recent analyses for the authenticated user
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
        "401":
          description: Not authenticated
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
...
    Resume:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        tag:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - content
        - tag
        - createdAt
        - updatedAt
    CreateResumeBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tag:
          type: ["string", "null"]
      required:
        - title
        - content
    UpdateResumeBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tag:
          type: ["string", "null"]
    Analysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        resumeId:
          type: ["integer", "null"]
        jobTitle:
          type: string
        companyName:
          type: ["string", "null"]
        jobDescription:
          type: string
        resumeContent:
          type: string
        matchScore:
          type: integer
        tailoredResume:
          type: string
        coverLetter:
          type: string
        keywordsMatched:
          type: array
          items:
            type: string
        keywordsMissing:
          type: array
          items:
            type: string
        interviewQuestions:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - resumeId
        - jobTitle
        - companyName
        - jobDescription
        - resumeContent
        - matchScore
        - tailoredResume
        - coverLetter
        - keywordsMatched
        - keywordsMissing
        - createdAt
    CreateAnalysisBody:
      type: object
      properties:
        resumeId:
          type: ["integer", "null"]
        jobTitle:
          type: string
        companyName:
          type: ["string", "null"]
        jobDescription:
          type: string
        resumeContent:
          type: string
      required:
        - jobTitle
        - jobDescription
        - resumeContent
    DashboardStats:
      type: object
      properties:
        totalResumes:
          type: integer
        totalAnalyses:
          type: integer
        averageMatchScore:
          type: integer
        plan:
          type: string
        usageCount:
          type: integer
        usageLimit:
          type: integer
      required:
        - totalResumes
        - totalAnalyses
        - averageMatchScore
        - plan
        - usageCount
        - usageLimit
...